Move PublicKeyTest code (for the abstract class static methods) out of Ed25519PublicKeyTest.
